terra-product-service/requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
Django==4.2.9
djangorestframework==3.16.1
drf-yasg==1.21.11
idna==3.11
inflection==0.5.1
mysqlclient==2.2.7
packaging==25.0
pika==1.3.2
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
sqlparse==0.5.3
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
gunicorn==21.2.0
django-cors-headers==4.3.1